Make the output reproducible
Whilst working on the Reproducible Builds effort [0] I noticed that
sphinx-panels and sphinx-design do not create reproducible output.

This is because it uses Python's uuid.uuid4 to generate unique identifiers, but
those numbers are random/nondeterminstic by design. This patch will seed these
random numbers from SOURCE_DATE_EPOCH if it exists, otherwise it will revert
back to random numbers.

I originally filed this in Debian as bug #1017475 [1], as well as within
sphinx-panels [2]

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/1017475
 [2] https://github.com/executablebooks/sphinx-panels/pull/82