feat: Sticky sync tabs and URL param
Add a Sphinx configuration option that is named
`sphinx_design_sync_tabs_storage_key` so that users
can set a project-specific local storage key rather
than risk clobbering the value for `sphinx-design-last-tab`
across different projects.

Add support for an optional URL search param of
like `?tab=blah` or `?code=blah` so that I can copy
and paste a URL and the receiver sees the tab
content that I see. By default, no param and you
get to add "Oh, and click blah" in your message.

If the URL has search param `?tab=blah`, set the
local storage value to `blah`. If `blah` matches
a synchronization key for a tab, show the tab.

When someone clicks a tab, set the local storage
value to the synchronization key value and also
update the URL search params with `?tab=<value>`
if the project configures a URL param.